refactor: rename providers to ports and adapters
