+++ __tests__/mocks/email.provider.mock.ts
@@ -1,6 +1,6 @@
 vitest.mock('src/infrastructure/providers/email.provider.js', async (importOriginal) => {
   const { ResendEmailProvider } =
-    await importOriginal<typeof import('src/infrastructure/providers/email.provider.js')>();
+    await importOriginal<typeof import('src/infrastructure/adapters/email.provider.js')>();
 
   class MockResendEmailProvider extends ResendEmailProvider {
     public override sendEmail = vitest.fn();
