+++ src/domain/ports/subscription.repository.ts
