+++ src/domain/ports/weather.provider.ts
