+++ src/domain/services/weather.service.ts
@@ -1,6 +1,6 @@
 import { Exception, ExceptionCode } from 'src/domain/entities/exception.entity.js';
-import { WeatherProvider } from 'src/domain/providers/weather.provider.js';
-import { Logger, LoggerProvider } from '../providers/logger.provider.js';
+import { WeatherProvider } from 'src/domain/ports/weather.provider.js';
+import { Logger, LoggerProvider } from '../ports/logger.provider.js';
 import { Weather } from '../entities/weather.entity.js';
 
 type Options = {
