+++ src/infrastructure/adapters/logger.provider.ts
@@ -1,5 +1,5 @@
 import { Logger as Pino, pino } from 'pino';
-import { Logger, LoggerProvider } from 'src/domain/providers/logger.provider.js';
+import { Logger, LoggerProvider } from 'src/domain/ports/logger.provider.js';
 
 type Options = {
   config: { PINO_LEVEL: string };
