+++ src/infrastructure/container.ts
@@ -1,20 +1,20 @@
-import { ConfigProvider } from './providers/config.provider.js';
-import { PinoLoggerProvider } from './providers/logger.provider.js';
-import { MetricsProvider } from './providers/metrics.provider.js';
-import { LoggerHttpProviderDecorator } from './providers/http/logger.provider.js';
-import { FetchHttpProvider } from './providers/http/fetch.provider.js';
-import { DbProvider } from './providers/db.provider.js';
-import { CacheProvider } from './providers/cache.provider.js';
-import { DrizzleSubscriptionRepository } from './repositories/subscription.repository.js';
-import { FastJwtProvider } from './providers/jwt.provider.js';
-import { ApiWeatherProvider } from './providers/weather/api-weather.provider.js';
-import { OpenMeteoProvider } from './providers/weather/open-meteo.provider.js';
-import { CacheWeatherProviderProxy } from './providers/weather/cache.provider.js';
+import { ConfigProvider } from './adapters/config.provider.js';
+import { PinoLoggerProvider } from './adapters/logger.provider.js';
+import { MetricsProvider } from './adapters/metrics.provider.js';
+import { LoggerHttpProviderDecorator } from './adapters/http/logger.provider.js';
+import { FetchHttpProvider } from './adapters/http/fetch.provider.js';
+import { DbProvider } from './adapters/db.provider.js';
+import { CacheProvider } from './adapters/cache.provider.js';
+import { DrizzleSubscriptionRepository } from './adapters/subscription.repository.js';
+import { FastJwtProvider } from './adapters/jwt.provider.js';
+import { ApiWeatherProvider } from './adapters/weather/api-weather.provider.js';
+import { OpenMeteoProvider } from './adapters/weather/open-meteo.provider.js';
+import { CacheWeatherProviderProxy } from './adapters/weather/cache.provider.js';
 import { WeatherService } from 'src/domain/services/weather.service.js';
-import { JsxTemplateProvider } from './providers/template.provider.js';
-import { ResendEmailProvider } from './providers/email.provider.js';
+import { JsxTemplateProvider } from './adapters/template.provider.js';
+import { ResendEmailProvider } from './adapters/email.provider.js';
 import { SubscriptionService } from 'src/domain/services/subscription.service.js';
-import { CronerCronProvider } from './providers/cron.provider.js';
+import { CronerCronProvider } from './adapters/cron.provider.js';
 import { Server } from './server.js';
 import { App } from './app.js';
