+++ src/infrastructure/controllers/metrics.controller.ts
@@ -1,5 +1,5 @@
 import { OpenAPIHono, createRoute } from '@hono/zod-openapi';
-import { MetricsProvider } from '../providers/metrics.provider.js';
+import { MetricsProvider } from '../adapters/metrics.provider.js';
 import { z } from 'zod';
 
 export const makeMetricsRoutes = (app: OpenAPIHono, metricsProvider: MetricsProvider) => {
